requirements.txt
torch
torchvision
tqdm
pyyaml
tensorboard
pillow
azureml-core
azureml-dataset-runtime
pandas
